Tidy up component list panel monolith
